test: move complete tests files without duplicating tests
