+++ tests/unit/array/oldproto/test_ravel_unravel.py
@@ -2,6 +2,8 @@ import numpy as np
 import paddle
 import pytest
 import tensorflow as tf
+import torch
+from scipy.sparse import csr_matrix, coo_matrix, bsr_matrix, csc_matrix
 
 from docarray import DocumentArray, Document
 from docarray.math.ndarray import to_numpy_array
@@ -11,8 +13,16 @@ def get_ndarrays_for_ravel():
     a = np.random.random([10, 3])
     a[a > 0.5] = 0
     return [
+        (a, False),
+        (torch.tensor(a), False),
         (tf.constant(a), False),
         (paddle.to_tensor(a), False),
+        (torch.tensor(a).to_sparse(), True),
+        # (tf.sparse.from_dense(a), True),
+        (csr_matrix(a), True),
+        (bsr_matrix(a), True),
+        (coo_matrix(a), True),
+        (csc_matrix(a), True),
     ]
 
 
@@ -52,13 +62,32 @@ def test_ravel_embeddings_tensors(ndarray_val, attr, is_sparse, da_cls):
     np.testing.assert_almost_equal(ndav, ndarray_val)
 
 
+@pytest.mark.parametrize('sparse_cls', [csr_matrix, csc_matrix, bsr_matrix, coo_matrix])
+def test_bsr_coo_unravel(sparse_cls):
+    a = np.random.random([10, 72])
+    a[a > 0.5] = 0
+
+    da = DocumentArray.empty(10)
+    for d, a_row in zip(da, a):
+        d.embedding = sparse_cls(a_row)
+
+    np.testing.assert_almost_equal(a, da.embeddings.todense())
+
+
 def get_ndarrays():
     a = np.random.random([10, 3])
     a[a > 0.5] = 0
     return [
-        (paddle.to_tensor(a), False),
+        (a, False),
+        (torch.tensor(a), False),
         (tf.constant(a), False),
+        (paddle.to_tensor(a), False),
+        (torch.tensor(a).to_sparse(), True),
         (tf.sparse.from_dense(a), True),
+        (csr_matrix(a), True),
+        (bsr_matrix(a), True),
+        (coo_matrix(a), True),
+        (csc_matrix(a), True),
     ]
