+++ tests/unit/math/oldproto/test_ndarray.py
@@ -2,12 +2,12 @@ import numpy as np
 import paddle
 import pytest
 import tensorflow as tf
+import torch
+from scipy.sparse import csr_matrix, coo_matrix, bsr_matrix, csc_matrix, issparse
 
-from scipy.sparse import issparse
-
+from docarray.math.ndarray import get_array_rows, check_arraylike_equality
 from docarray.proto.docarray_pb2 import NdArrayProto
 from docarray.proto.io import flush_ndarray, read_ndarray
-from docarray.math.ndarray import get_array_rows, check_arraylike_equality
 
 
 @pytest.mark.parametrize(
@@ -22,8 +22,15 @@ from docarray.math.ndarray import get_array_rows, check_arraylike_equality
 @pytest.mark.parametrize(
     'arraytype',
     [
-        paddle.to_tensor,
+        list,
+        torch.tensor,
         tf.constant,
+        paddle.to_tensor,
+        torch.tensor,
+        csr_matrix,
+        bsr_matrix,
+        coo_matrix,
+        csc_matrix,
     ],
 )
 @pytest.mark.parametrize('ndarray_type', ['list', 'numpy'])
@@ -50,8 +57,16 @@ def get_ndarrays():
     a = np.random.random([10, 3])
     a[a > 0.5] = 0
     return [
-        paddle.to_tensor(a),
+        a,
+        a.tolist(),
+        torch.tensor(a),
         tf.constant(a),
+        paddle.to_tensor(a),
+        torch.tensor(a).to_sparse(),
+        csr_matrix(a),
+        bsr_matrix(a),
+        coo_matrix(a),
+        csc_matrix(a),
     ]
